backend/fleet-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.2.2
sqlalchemy[asyncio]==2.0.49
asyncpg==0.31.0
redis==5.0.1
